feat:#103 관리자 페이지 조회 로직 완료

+++ 1_CodeSoruce/src/main/java/com/bughunters/mountainspirit/admin/query/dto/MemberInfoStatsDTO.java
@@ -11,7 +11,7 @@ import java.util.Map;
 @NoArgsConstructor
 @AllArgsConstructor
 public class MemberInfoStatsDTO {
-    private int totalMembers;                              // 전체 회원 수
+    private Long totalMembers;                              // 전체 회원 수
     private List<GenderRatioDTO> genderRatio;         // 성별 비율
     private List<AgeDistributionDTO> ageDistribution;    // 연령별 회원 분포
     private List<MonthlyNewMemberDTO> monthlyNewMembers;  // 최근 1년 월별 신규 가입자
